ani: add anidb.app streaming provider alongside AnimeTosho
A provider selector (config "source": "torrent"|"anidb") switches the
backend. anidb.app streams directly (mpv + HLS URL) instead of torrenting
(webtorrent + magnet). The release picker is reused for both — for anidb
its rows are audio×resolution variants (group filter = sub/dub, quality
filter = resolution).

Architecture:
- internal/playable/ (NEW): provider-agnostic Release type. Both providers
  produce *playable.Release. Torrent fields (Magnet/Seeders/Size) + stream
  field (StreamURL) on one type; IsStream() branches play + render.
- internal/anidb/ (NEW): anidb.app client. Search (/browse HTML), Episodes
  (JSON), FetchReleases (languages → embed → master.m3u8 → variants).
  Handles cumulative episode numbering via per-show offset
  (EpisodeOffset). AiredCount returns the latest available MAL-equivalent
  episode (float64 — fractional specials like 3.5 are preserved).

Episode mapping:
- anidb uses cumulative numbering across seasons (e.g. Slime S4 = eps 73–88).
  offset = first_real_episode − 1; MAL_ep ↔ anidb_ep = MAL_ep + offset.
- The release picker defaults to watched+1 (next unwatched) when it exists
  on anidb; falls back to the latest available otherwise. The default is
  recomputed each play-loop iteration so it advances after watching.
- MAL write-back writes the correct per-season episode number
  (anidb_ep − offset).

Config:
- Per-provider group/quality: "group"/"quality" for torrents,
  "anidb_group"/"anidb_quality" for anidb. Sort is shared.
- SaveFilters is source-aware.

Player: RunPlayURL (mpv + stream URL, no webtorrent).

Title search: cleanQuery only cuts at " — " (not ":" which is part of
titles like "Re:Zero").

Phase 0 (Cloudflare) + Phase 0.5 (episodes/aired/stream probes) passed
with plain net/http — no browser-fingerprint dependency needed.

+++ internal/tui/anime_picker.go
@@ -399,13 +399,13 @@ type animePicker struct {
 	// disables the "watched/aired/total" display). aired is a session-scoped
 	// *AiredCache: each anime's count is computed at most once per session (incl.
 	// 0/failed — no retry), shared across pickers via app.Run.
-	latestEpisode func(item *mal.Item) int
+	latestEpisode func(item *mal.Item) float64
 	aired         *AiredCache
 
 	// latestEpisodePrefetch is the background prefetch variant (fast-only: no
 	// Jikan, skips aid-unresolved items). nil ⇒ covers are still paged but no
 	// aired prefetch.
-	latestEpisodePrefetch func(item *mal.Item) int
+	latestEpisodePrefetch func(item *mal.Item) float64
 
 	// prefetchSem bounds concurrent aired-episode fetches. Animetosho times out
 	// (~30s) when too many hit it at once — a full season has ~90 airing anime and
@@ -458,7 +458,7 @@ func animeCacheKey(source AnimeSource, query, season string) string {
 	return fmt.Sprintf("%d|%s|%s", source, query, season)
 }
 
-func newAnimePicker(source AnimeSource, query string, load AnimeLoad, applyStatus func(int, int, StatusAction) bool, applyScore func(int, int) bool, applyWatched func(int, int) bool, latestEpisode func(*mal.Item) int, latestEpisodePrefetch func(*mal.Item) int, debug bool) *animePicker {
+func newAnimePicker(source AnimeSource, query string, load AnimeLoad, applyStatus func(int, int, StatusAction) bool, applyScore func(int, int) bool, applyWatched func(int, int) bool, latestEpisode func(*mal.Item) float64, latestEpisodePrefetch func(*mal.Item) float64, debug bool) *animePicker {
 	y, s, label := mal.CurrentSeason()
 	ap := &animePicker{
 		source:                source,
@@ -933,7 +933,7 @@ func (m *animePicker) handleKey(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
 		if it := m.currentItemCopy(); it != nil {
 			// Carry the cached aired count so the release picker reuses it
 			// instead of re-fetching on entry.
-			it.AiredEps = m.aired.value(it.MalID)
+			it.AiredEps = int(m.aired.value(it.MalID))
 			m.result.Anime = it
 			return m, tea.Batch(tea.Quit, m.quitCmd())
 		}
